agent ai/requirements.txt
langchain==1.2.14
langchain-core==1.2.25
langchain-openai==1.1.12
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
mysql-connector-python==9.6.0
openai==2.30.0
pydantic==2.12.5
python-dotenv==1.2.2
sqlalchemy==2.0.48
pymysql==1.1.2
